+++ Source/SourceKittenFrameworkTests/ClangTranslationUnitTests.swift
@@ -31,8 +31,17 @@ class ClangTranslationUnitTests: XCTestCase {
         let headerFiles = [fixturesDirectory + "Musician.h"]
         let compilerArguments = ["-x", "objective-c", "-isysroot", sdkPath()]
         let tu = ClangTranslationUnit(headerFiles: headerFiles, compilerArguments: compilerArguments)
-        let comparisonString = (tu.description + "\n").stringByReplacingOccurrencesOfString(fixturesDirectory, withString: "")
-        let expectedOutput = File(path: fixturesDirectory + "Musician.xml")!.contents as String
-        XCTAssertEqual(comparisonString, expectedOutput, "Objective-C docs should match expected ouput")
+        let escapedFixturesDirectory = fixturesDirectory.stringByReplacingOccurrencesOfString("/", withString: "\\/")
+        let comparisonString = (tu.description + "\n").stringByReplacingOccurrencesOfString(escapedFixturesDirectory, withString: "")
+        compareJSONStringWithFixturesName("Musician", jsonString: comparisonString)
+    }
+
+    func testRealmObjectiveCDocs() {
+        let headerFiles = [fixturesDirectory + "/Realm/Realm.h"]
+        let compilerArguments = ["-x", "objective-c", "-isysroot", sdkPath(), "-I", fixturesDirectory]
+        let tu = ClangTranslationUnit(headerFiles: headerFiles, compilerArguments: compilerArguments)
+        let escapedFixturesDirectory = fixturesDirectory.stringByReplacingOccurrencesOfString("/", withString: "\\/")
+        let comparisonString = (tu.description + "\n").stringByReplacingOccurrencesOfString(escapedFixturesDirectory, withString: "")
+        compareJSONStringWithFixturesName("Realm", jsonString: comparisonString)
     }
 }